Silently drop messages that match the blacklist instead of bouncing it back to sender.  This was done to fix the case where a TMDA user adds someone to her blacklist who regularly posts to a mailing list she is part of.  If the list is filtered through TMDA, the blacklist bounce could go to either:
a) the blacklisted user, who will now be harassed with a bounce
message every time she posts to the mailing list.

b) the list VERP, which might think that the user's mail is bouncing
and unsubscribe or otherwise harass the user.

c) the list owner/admin/nazi's address, who may get annoyed and remove
the user from the list and/or publically chastise her.

d) in the case of extremely badly-designed list software, TMDA might
even wind up bouncing the mail to the entire mailing list which could
cause all sorts of chaos.

None of these cases is appropriate.