For a message sent to local@domain, confirm-*.HMAC is now just appended to local so that when the confirmed message is released, it will be processed by the same dot-qmail file that called tmda-filter in the first place.
Code was also added to use the address that the message was originally
sent to (original envelope recipient) as a base instead of RECIPIENT
from the environment (which is munged under a qmail virtualdomain).
Unfortunately, qmail provides no easy method of obtaining this
information, so the control/virtualdomains file has to be parsed.

The templates were also modified to send confirmation messages with a
"From:" header that matches the original recipient for added
flexibility.