Fix a bug in the Postfix deliver method where the exit code from LOCAL_DELIVERY_AGENT was not being propagated to the MTA.  (e.g, tmda-filter returns 0 if procmail returns 77.)  The exit status of the pipe's close() function has the signal the command died from in the low byte (if any) and the exit status of the command in the high byte.
The fix is to do some more processing on the exitcode before passing
it to sys.exit.

What I've done here is check for unexpected death by signal in the lda
and deferred when that happens.  I tested this by setting
LOCAL_DELIVERY_AGENT to a small C program that segfaults.  It seems to
work.

Otherwise, I examine the exit status.  If it is one postfix will
understand (64 to 78 inclusive), then we honor it and return that
value.

Otherwise we defer.  I think it is useful to defer when the exit code is
not 64 to 78 inclusive.  It will save people's butts when their LDA
isn't acting correctly.  maildrop and postfix both seem to return 75
when they encounter a random error, but other home-grown LDAs won't.

Patch contributed by Matt Armstrong.