Fix a bug in tmda-inject reported by Drew Raines in <20011115114238.D11178@phg.mc.vanderbilt.edu> on tmda-users.
When the MUA "bounces" a message, tmda-inject was modifying the
existing From: header which was incorrect.

The problem was that tmda-inject was not differentiating between
sending a message for the first time and resending a message
(bouncing).  In each case, it grabbed the From: header and uses that
as the base for tagging both the From: header and the envelope sender
address.

For a bounce, tmda-inject should not touch the existing From: header
at all, and should use Resent-From: (not From:) as the base for the
tagged address.

The solution is to first detect whether we are sending or resending by
checking for a `Return-Path' header, and then:

if sending an original message:
   - derive the tagged address from the existing "From:" header
   - tag both "From" and the envelope sender
   - the message's "From" address and envelope sender will match

if resending a message:
   - derive the tagged address from the existing "Resent-From:" header
   - tag both "Resent-From" and the envelope sender (leaving From: intact)
   - the message's "Resent-From" address and envelope sender will match