Add an option to the `sender' action.  When sender=address, address is used the create the sender cookie instead of the message recipient's address.
This is useful for example when posting to a subscriber-only mailing
list to which you are subscribed with a sender address.  e.g, in
FILTER_OUTGOING:

# this posts with a sender address that doesn't match my subscribed
# address
to mailing-list@domain.dom sender

# instead we use this to post with owner-mailing-list@domain.dom's
# sender address, which does match
to mailing-list@domain.dom sender=owner-mailing-list@domain.dom

Previously, you would have needed to use `ext' or `exp' to accomplish
this less elegantly.