New feature.  If the message has an empty envelope sender (i.e, SENDER=), change it to the string `<>'.  This provides a way of matching empty envelope senders in the filter files.
This provides a way of dealing with the dilemma where if you send an
untagged message and it bounces, tmda-filter will send a confirmation
back to <> which is pointless.  You can now train your incoming filter
to accept all bounces (messages with an empty envelope sender).

This also distinguishes the empty envelope sender in the LOGFILE.

Requested by Martijn Koster.