New feature.  Optionally add an `X-TMDA-Fingerprint' header to outgoing client-side messages.  This fingerprint is an HMAC of one or more user-specified headers from the message.
This can be used to uniquely identify the message, and the keying
insures that the fingerprint was generated by you.

One application for this might be an "auto-ack" system used to
automatically reply to qsecretary or TMDA confirmation requests.  If
your outgoing messages contain X-TMDA-Fingerprint, you can check for it
later when qsecretary has returned the message to you in a
confirmation request.  After validating the fingerprint, your program
automatically sends back a reply, and your original message is confirmed.