Update the Sendmail documentation.
Clarify the required format for SENDER, RECIPIENT, and EXTENSION for
the Sendmail user who wants to trying setting these variables in a
manner other than that described in the documentation.

Also, change the sample .procmailrc to strip angle brackets <> in
addition to leading whitespace from the X-Envelope* headers.  Brian
Lalor tells us this is necessary because angle brackets are added when
mail it sent from a remote host.