Bugfix.  Keyword addresses were broken under a qmail virtualdomain. The problem was that the "cookie value" for a keyword address is the value of the EXT environment variable.  However, this isn't accurate under a qmail virtualdomain as EXT also includes the virtualdomain "prepend" as well as the virtual username.
The solution is to strip both of these off the front of
address_extension at the same time we are fixing up recipient_address.

The fact that qmail leaves no trace of the _original_ envelope
recipient in the environment under a virtualdomain is again, a design
flaw in my opinion, and the reason for all this kludgy parsing.
Grrr..

Problem reported by Andy Bradford in
<200112312303.QAA20596@sicilia.bradfordfamily.org> on tmda-users.