Internal reorganization/bugfix.  Previously, when tmda-filter received a user-confirm-done.HMAC address, it would only verify the HMAC before allowing the message through.  This was problematic in that if someone got ahold of a valid user-confirm-done.HMAC address (e.g, from the TMDA logfile, forwarded message headers, etc.) they could sneak messages through TMDA with that address.  This was an accident waiting to happen.
Now, we also check whether the message associated with the
'confirm-done' address is actually in the pending queue -- if it
isn't, that means the message was already delivered, and someone is
playing silly buggers, so we bounce it.  If it is, this indicates a
valid (first time) confirmation, and we accept the message, and then
delete its counterpart from the pending queue.

Some of the functions in tmda-filter were reorganized to accommodate
this change.