Internal reorganization: The current tmda-filter is moved into tmda-rfilter, and tmda-filter is now a simple wrapper script.
This wrapper will be called from the .forward/.qmail/.procmailrc
instead of calling tmda-rfilter directly.  It's purpose is to load the
"real" tmda-filter, catching any exception that isn't the SystemExit
exception, logging it to ~/TMDA_DELIVERY_FAILURE, and then deferring
the delivery.  If it can't log to the file, the traceback is printed
to stdout so at least it will be included in the eventual bounce
message.

This way, no matter how badly tmda-rfilter fails, it will still get
its error logged and the mail will be properly deferred.  This will
provide an MTA-independent error logging method, and make debugging
problems much easier.

We can't rely on the command output being printed to syslog, as 1) not
all MTAs do this, and 2) the TMDA user may not have permission to read
his /var/log/maillog.  What's more, exim unconditionally bounces if
the delivery command writes to stderr, and uncaught python exceptions
write to stderr.

Suggested by Matt Armstrong in <871yhga95i.fsf@squeaker.lickey.com> on
tmda-workers.