tmda-check-address now reports that it only works for 'dated' or 'sender' addresses as the 'keyword' address checking is currently broken.
The problem is the address validation code is out of sync
with tmda-rfilter, and code duplication makes this a constant
struggle, so we punt for the time being.

tmda-check-address needs to be rewritten.  The consensus is to
consolidate the tagged address manipulation code into a common Python
class that represents an address.  This will help centralize address
manipulations needed by all TMDA programs.