For file based sources (from-file, from-cdb, etc.) we previously just passed over the associated line in the filter file if either the file source was missing, or if an exception occured when trying to process the file.
I now think was this incorrect behavior, as no indication was given as
to why the match failed.  I feel it's better to defer the message and
give folks a chance to correct the problem so the message is processed
the way they initially intended.

So now we try to open and process the file unconditionally, and let
any exceptions be caught by the tmda-filter wrapper.