Introduce tmda-pending, a pending queue manipulation tool.  This long overdue tool should hopefully help the confidence of newbies who are reluctant to turn their mail over to TMDA due to fears of legitimate mail being lost in pending.
It comes at a time when one of the most FAQ on tmda-users is how to
force a message out of the pending queue which there is no consistent
way to do across MTAs.  tmda-pending provides a universal way to
release messages as it works in the same way that pending messages are
released when someone replies to a confirmation request.

Several other features are included such as the ability to delete
messages, and print summaries with release address links.

Both interactive and batch processing modes are available.