Introduce local mail delivery capability.
This was done in response to several user requests for the ability to
specify various delivery instructions from the incoming filter file,
including delivery to a program, delivery to a Maildir or mbox, and
forward to an e-mail address.

The delivery instruction is given as an option to the existing
ok,accept,deliver action.

The local delivery work is done in the new Deliver module.  Initially,
'forward' and 'program' instructions are supported.

NOTE: qmail users must now invoke tmda-filter with 'preline' in their
dot-qmail.  e.g,

|preline /usr/local/bin/tmda-filter