Bugfix.  TMDA was raising exceptions when a program (such as dmail) exits 0 but outputs to stderr.  We fix that, as well as an exit-on-signal problem.
Reported by Tim Legant in <86it8ejfxs.fsf@skitty.catseye.net> on
tmda-workers.

This fix may potentially pose problems if the message is piped to
procmail, and procmail fails yet still returns 0, which it allegedly
does.  This is broken behavior, and I'd rather not support it by
coding around it.  This would also make life difficult for users of
correctly behaving programs such as dmail, requiring them to
specifically send stderr output to /dev/null.

The needs of the many outweigh the needs of the few who use braindead
software.