Bugfix.  tmda-pending was skipping over delivered messages in all cases, which means they would never be removed from disk.
Now, we don't skip them if running in batch + delete mode (-bd) so
they can be properly purged along with old pending messages.

There were some other suggestions on tmda-workers which I couldn't
make up my mind on, so for now we will be conservative.  The main
purpose of tmda-pending is to operate on pending and not delivered
messages anyway, hence its name.