Bugfix.  This fixes a nasty, potentially mail losing bug that affects non-qmail users only.  This bug has been present for 4 months now, and I'm simply amazed no one has noticed and/or reported it..
When mail is sent to a 'confirm-done' address, I had been checking to
make sure the message was not previously delivered before accepting
it.  It if bad been, the message was dropped.  This was done so that
someone couldn't sneak mail through simply by coming across an old
'confirm-done' address (perhaps from your unprotected TMDA logfile).

The problem is that if the final delivery was deferred, perhaps due
to an error in DELIVERY (previously LOCAL_DELIVERY_AGENT), the next
time around, TMDA would erroneously think that this message had
already been delivered, when it actually hadn't, and the message would
hit the floor.

Ideally, we'd only record the message as delivered *after* the
delivery method had returned, but we can't do this.  If we are running
qmail, delivery may be an exit(0), meaning program control will never
return, and we'd never get the change to record the delivery.

The chance that someone will find a 'confirm-done' address and start
spamming it is slim, but nonetheless I don't want TMDA to have any
holes, no matter how small.  I need to find some other way of making
sure only TMDA can send mail to a 'confirm-done' address.