Change of mode for when tmda-ofmipd is running as `tofmipd'. The problem was that since the process was running as user `tofmipd', users could put code in their .tmda/config or in a deliver instruction that read other user's crypt_key files.
We need a way to assume the identity of the TMDA user so that this is
not a concern.

The daemon now uses seteuid() to run as tofmipd during the SMTP AUTH
session and then do a little footwork -- switch back to root, spawn a
new process that does setuid() to the authenticated user and handles
delivery, then switch tmda-ofmipd back to user tofmipd with seteuid().

It's still not clear whether this strategy is secure or not and
therefore whether it will be retained.  The daemon runs with EUID
tofmipd, but with UID root.