Bugfix.  When releasing a message from pending, it isn't always enough to just remove the first Delivered-To line.  If users are forwarding the message around, this won't prevent a mail loop.
Now we find the first Delivered-To line containing the same recipient
as in X-TMDA-Recipient, and remove that and any Delivered-To lines
above it.  Usually, that'll be just the first Delivered-To, but in the
case of forwards using the RECIPIENT_HEADER mechanism, we'll take out
any forwards and let the whole sequence get built again (during the
re-injection).

Bug reported by Tim Legant in <86d6ugza2c.fsf@skitty.catseye.net>
on tmda-users.