Add some new wording to emphasize that TMDA has features (e.g, `dated' addresses) that greatly reduce the number of legitimate senders who are actually faced with a confirmation request.
The biggest misconception about TMDA is still that it's nothing more
than a "whitelist checker", and that it shifts the burden of the SPAM
problem onto the poor innocent sender.  While this is true to some
extent, using TMDA in its full form will reduce this burden
significantly.  As pointed out in the FAQ, over the past year and a
half, only ~6% of my legitimate senders have actually had to reply to
a confirmation request.  These are generally folks e-mailing me out of
the blue, or replying to an old USENET or mailing list post of mine.
That's very little time wasted, even if it is someone else's time.

I'm also tired of chasing down the misconceptions and erroneous
gibberish propagated by all the knuckle-heads out there who like to
make assumptions without reading documentation.  Hopefully word of
mouth will be loud enough to quelch most of the ignorance.