Add Barry Warsaw's `email' package version 2.3 to the distribution. email is covered by the "good" Python license which is compatible with the GPL, so this is all kosher.
email is a new Python package for parsing, handling, and generating
email messages and other RFC 2822 style documents.  It is intended to
replace several older modules in the standard distribution, such as
rfc822, mimetools, multifile, mimify, and MIMEWriter, and such
non-standard modules as mimecntl.

Using email within TMDA should provide additional performance,
robustness, and make new features such as MIME and multi-lingual
support much easier to implement.

We include the package so we don't have to require a minimum of Python
2.3 in order to run TMDA (which is too restrictive at this point).
Instead, only a minimum of Python 2.1 will be required.  This
flexibility comes at the expense of adding an additional 144K to TMDA.

NOTE to developers:

The email package is designed to be a top level package.  So, pure
Python programs like TMDA should always import through "email" as the
top level package.

This means we can't do relative imports from the email package.  e.g,
we must use:

from email import Parser

rather than:

from TMDA.pythonlib.email import Parser

The latter will not work.

email is installed underneath the TMDA module directory:

TMDA/
    pythonlib/
             email/

But we still must add TMDA/pythonlib to the front of sys.path
regardless of whether we are running TMDA from a self-contained
directory, or a package installation like an RPM.

Accordingly, code was added to each bin/tmda-foo program (for package
installation) as well as to bin/paths.py (for directory installation).