New feature.
TEMPLATE_DIR_MATCH_SENDER provides a way to further specialize the
template selection process.

When enabled, TMDA looks for a subdirectory of TEMPLATE_DIR that
matches the sender address, and then increasingly general portions of
the domain part of the address.

For example, if mail arrives from foo@bar.baz.de, we'd look for
templates in these subdirectories under TEMPLATE_DIR, in this order:

  foo@bar.baz.de/
  bar.baz.de/
  baz.de/
  de/

If no sender based templates are found, TEMPLATE_DIR and then the
default template locations are tried.