Bug workaround. If in your template, you include the original subject in your Subject line, e.g:
  Subject.US-ASCII: Please confirm your message (Re: %(subject)s)

And the incoming message has an encoded subject that your Python
can't decode (e.g, BIG5, EUC-KR), a LookupError will be raised.

As a temporary workaround, now we don't attempt any decoding if
the specified charset for Subject is US-ASCII.

This means that users who use a non-ASCII Subject field in their
template, _and_ also include the original subject will still have
this problem as decoding will be attempted. e.g,

  Subject.LATIN-1: Bitte Best�tigen (Re: %(subject)s)

This patch covers the majority of users though, and buys us some time
until a real solution can be found. Unfortunately, this is a rather
difficult problem to solve. For details on why, see the thread
starting at:

  http://article.gmane.org/gmane.comp.python.mime.devel/342

Until (if?) we fix this, users who report this problem will have a few
options to choose from:

1) Use only US-ASCII in the Subject field.

2) Remove %(subject)s from the Subject field.

3) Install ChineseCodecs, JapaneseCodecs, and KoreanCodecs.

4) Once the above Codecs ship with Python, upgrade to that version of
   Python.

Hmm, perhaps it's not the end of the world if we can't figure this one
out, as there are still some options for the user.

This bug was reported by Martin Klaffenboeck in
<20021108163353.GC810@martin.kdrache.org> on tmda-users.