Moved crypt_key loading code to the end of the program so that other dependent variables will already be set by the time it happens.
Added a PENDING_FILE_PERM so that users can override file permissions on pending
e-mail files.  (This was needed for no-su mode of tmda-cgi.)

Modified crypt_key loading code so that if PENDING_FILE_PERM is anything other
than 0600, that errors will be ignored during the loading process.  These two
entities may not seem to be related, but they are.  If you need permissions on
your pending files other than 0600, then you may be running in no-su mode (for
tmda-cgi).  In no-su mode, the CGI will still need to have defaults set and
loaded from configuration files, but the crypt_key's will not be accessible.  It
is important that this code not throw an exception when imported or the import
will fail.