Add CGI_URL for things like constructing the message release URL when using tmda-cgi.
If this variable is set, tmda-rfilter constructs a release URL
for the message and stores it in `confirm_accept_url'.

Then the user can imbed the URL in her template by adding something
like this to confirm_request.txt:

  To release your message, click on %(confirm_accept_url)s

An example URL is:

http://www.domain.tld/tmda.cgi?684.1038959858.48460.c7a2e8

The first integer is the euid which is necessary for the cgi to know
which pending directory to look in, which configuration file to load,
etc.

The next two portions identify the message-id on disk, and the last is
the HMAC. These three are the same as the "cookie" included within the
confirmation e-mail address.