The Manager is a kind of a shell to control TMDA. For the moment, it can only do what tmda-gui does, ie. managing pending queue and generating/checking addresses.
It's meant to run as a shell, or as a daemon. The latter mode enables a
user to use tmda-gui from a remote location, with no need of the TMDA
modules.

tmda-manager uses a new module called Auth.py, based on the code from
tmda-ofmipd. At some point, we'll need to move the common code into
Auth.

This is an early alpha release, so be careful with it. Try it on test
users first.

tmda-manager syntax:

Authentication:
 auth user pass

List pending msgids:
 pending

Handle message:
 message action msgid
    where action is one of show, terse, delete, release, whitelist,
        blacklist
    and msgid is the message id as shown by the pending command.

Generate address:
 address type option
    where type is one of dated, sender, keyword
    and option is the date, the sender address or the keyword.

Check addresses:
 checkaddress address arg
    where address is the address to check and arg is the sender address
        if needed.


More on this later.