Add our own message module which inherits from email.Message.  We do this to change certain default behaviors such as "From" with ">".
It's not the job of the MUA to escape From, so we change this in
tmda-inject.

This is a half-assed fix, because we need to do this to the
tmda-filter side as well, since escaping only needs to be done when
delivering to mbox files. However, this has proven more difficult than
anticipated, so it will have to wait.