Added a mechanism to report errors in a comment portion of the login HTML.
Changed the session mechanism to keep two files, one with just enough
information to validate a session number and associate it with a user, and one
with all the persistent settings a user might need.

Changed the name of Pending.py to PendList.py to keep it from being mistaken
for TMDA's Pending.py.

Moved session settings from config to compiled in.

For security reasons, changes to the user's UID are now done with a setuid
instead of a seteuid.  This will prevent the user from doing a seteuid(0) in
their config file and executing commands as root (in system-wide mode).