Added more debugging code to Auth.py.
Removed virtual user processing from Util.getuserparams().  The code that was
in there would ONLY work if run as root.  If we need to look up user details on
a virtual user as anyone but root, we have to do this in a very different way.

Added Util.RunTask() which correctly fetches the data we need for virtual user
processing (see above).  Unfortunately, this code returns raw data that still
needs to be processed by a stub program.

For the time being, the stub code has NOT been added to the TMDA project.
Instead, I'm putting the one stub I have directly into a tmda-cgi directory.
I hope that the stub(s) will eventually become part of TMDA, but before it does
I think that it needs to:
1. stabilize (been making lots of changes to it lately)
2. be accepted by the community.