Added code for local config editing.
Added a "NoOverride" section to defaults.ini for settings that should only be
in defaults and never overridden by a theme.

Grey out the first/prev and/or next/last if we are at the beginning and/or end
of the message list.

Moved variable General:UserSplit to NoOverride:UserSplit.

Added some stuff left out of the local config template.