Replace use of Message.as_string() with our own implementation that creates an email.Generator.Generator object and calls .flatten(msg) on it.
The problem with Message.as_string() is that it returns a textual
representation of the message with all leading "From" lines escaped,
and with all header lines wrapped at 78 characters.  This is a
conservative approach, but isn't appropriate for our use:

1) "From" line escaping is only appropriate when writing mbox files,
    a very specific case.

2) Wrapping header lines is only appropriate when we are sending
   outgoing mail, not for example when we are delivering incoming mail.

We need more fine-grained control over this behavior, so
Message.as_string() is useless.

Our implementation by default creates text without escaping From lines
and without wrapping any headers.  This is the most common behavior in
TMDA.  Then we selectively make exceptions when necessary.

TMDA should now Do The Right Thing.