Clarify the pre-requirements.  Too many knuckleheads make the assumption that TMDA is a desktop product and download before reading the requirements page.
I'm also tired of people asking about a Windoze version
of TMDA, so I've tried to make that issue abundantly clear
as well.

The printing of the Sunday Times article about TMDA was
the straw that broke the camel's back, as I probably
received 2-dozen of these questions today alone!