New feature: automated pending queue cleanup.  Upon receipt of an incoming message, tmda-filter will trigger a cleanup of old messages in the pending queue, dependent on PENDING_CLEANUP_ODDS.  The lifetime of a message is dictated by PENDING_LIFETIME.
Since this isn't run on every incoming message, performance impact
should be minimal, and offers a more convenient alternative to
running a cron job or cleaning messages by hand.  This inconvenience
is exacerbated by large and/or multi-user installations such as
under vpopmail/tmda-cgi.

It's also non-intuitive that messages will live in the pending queue
indefinitely unless the user intervenes, and in-fact this was one of
the most FAQ on tmda-users.