Bugfix/security.
It looks like this week's TMDA flame-war on the qmail list did have
one positive outcome, and that is a new method of invoking
/usr/sbin/sendmail that doesn't use the shell.  Thus we avoid the
possible security implications as well as argument quoting and
escaping issues.

You can avoid the shell by passing a tuple of arguments to
Util.pipecmd() as the command instead of a string.  This will cause
the popen2.Popen3() code to execvp() "/usr/bin/sendmail" with these
arguments exactly, with no trip through any shell.

We already knew this, as ancient TMDA did exactly this.  The problem
was that we kept getting burned by silent failures in the case of a
non-zero exit from the child process (e.g, if /usr/sbin/sendmail was a
truncated file, the message would silently be dropped on the floor).
What I didn't realize is that I needed to call wait(), capturing the
return value, and then use os.WIFEXITED(), os.WEXITSTATUS(), and
friends to get the exit status of the invoked command.

Our Util.pipecmd() function already does the above, so we just
had to pass the arguments in as a tuple, and now we get the best
of both worlds.  The scary security warning has been removed
from the Defaults.OUTGOINGMAIL comments.

Thanks to Chuck Cazabon (private e-mail xchange) for helping me iron
this out.