Change the default value for OUTGOINGMAIL to 'sendmail', and also rename OUTGOINGMAIL to MAIL_TRANSPORT.  Rationale:
I made SMTP transport the default for OUTGOINGMAIL in the past for two
reasons:

1) Since our /usr/sbin/sendmail implementation went through the shell,
   there was a potential security risk with incoming mail.  If someone
   embedded the right shell commands in their Return-Path, they could
   get executed by TMDA inadvertently.

2) Argument quoting and escaping issues when TMDA tried to respond to
   a really malformed address that confused the shell.

However, the new implementation doesn't use the shell at all, so the
above problems should hopefully be non-issues.

In addition, there were the following problems with the SMTP
default:

a) qmail doesn't let localhost relay by default, so we get 1001
   questions on tmda-users first about why TMDA isn't sending any
   mail, and then about how to setup qmail's obscure relaying
   mechanism.  This is FAQ 3.3.  The user probably just ends up
   setting OUTGOINGMAIL to 'sendmail' to fix things.

b) With Sendmail/Postfix/Exim, if TMDA tries to respond to a bogus
   local address, the MTA refuses the transmission during the SMTP
   transaction resulting in an LOGFILE_DEBUG traceback.  This is FAQ
   3.7 essentially.  The user probably ends up setting OUTGOINGMAIL to
   'sendmail' to fix things.

Reversing the default should solve both of these issues, with no
additional problems created.

The name "OUTGOINGMAIL" proved non-intuitive and confusing, so
it is renamed to something less ambiguous.