Bugfix.  tmda-ofmipd was not able to send mail to addresses which contained apostrophes (e.g, sinead.o'connor@domain.dom) because it was escaping them with \\.
This was originally done to to prevent shell exploits when the
tmda-inject command-line was expanded.  We can now avoid the shell
completely by passing a sequence to Util.pipecmd() instead of a
string, and thus we no longer need to escape the apostrophes.

Thanks to David Collantes for reporting this on tmda-users.