Bugfix.  If AUTORESPONSE_INCLUDE_SENDER_COPY = 1 and the message we are responding to lacked a charset parameter (usually in Content-Type), the resulting text/rfc822-headers bodypart would have no charset parameter itself.  e.g,
  MIME-Version: 1.0
  Content-Type: text/rfc822-headers; charset
                                     ^^^^^^^

We now pass the default charset (US-ASCII) in as a fallback charset
for these situations.

Also, add a Content-Disposition header field as described in RFC 2183
to the main message headers as well as each MIME bodypart when
auto-responding.

The `inline' disposition is used in all cases which tells the MUA that
each bodypart should be displayed inline automatically upon display of
the message, rather than treated as an attachment.

Content-Disposition is an optional header field, and in its absence,
the RFC says the MUA may use ``whatever presentation method it deems
suitable.''  In most cases, the MUA will likely do the right thing
with a TMDA auto-response, but this just makes sure it does.  If the
MUA incorrectly treated a bodypart as an `attachment', the
recipient might not see the confirmation request leading to a lost
message.