Bugfix.  When AutoResponse.record() tries to write an entry to ~/.tmda/responses/, an IOError (errno.ENAMETOOLONG) is raised if the filename exceeds the maximum number of chars allowed in a filename (255 max on POSIX systems).
This came to light when a user received a message with an enormously
long envelope sender address.  Surprisingly, rfc 2821 does allow such
lengths (section 4.5.3.1).  It only has recommendations on maximum
lengths, and gives the SMTP server the option to refuse, but does not
require it.

To fix, we truncate the sender address in Util.normalize_sender() at
233 chars.  Timestamp is a long, max 10 characters; PID might be a
long on 64-bit architectures, so max 10 characters; 2 separator dots
('.') are 2 characters.  Therefore, the email address can be a maximum
of 255 - 22 or 233 characters so that the entire filename is <= 255.

Will apply to main trunk as well.