New feature.
Inspired by qmail-inject's QMAILMFTFILE feature, MAIL_FOLLOWUP_TO
automatically adds (via tmda-inject) a Mail-Followup-To field for
messages sent to mailing lists.

MAIL_FOLLOWUP_TO can either be a list of mailing list addresses, or a
string pointing to a file containing mailing list addresses, one per
line.  If To or Cc in the message includes one of those addresses
(without regard to case), tmda-inject adds a Mail-Followup-To field
with all the To+Cc addresses.  tmda-inject does not add
Mail-Followup-To to a message that already has one, or if the address
file does not exist.