Attempt to solve the discrepency of '' and '<>' between the various MTAs for sending mail with a null envelope sender.  The 4 supported MTAs do not have the same behavior for doing so which you can test with:
  $ echo | /usr/sbin/sendmail -i -f "<>" -- you@yourdomain.dom
  $ echo | /usr/sbin/sendmail -i -f "" -- you@yourdomain.dom

Postfix/qmail require an empty string ''.  If you attempt to use '<>',
they will not complain, but the resulting envelope sender will not be
correct. e.g, under qmail:

  Return-Path: <"<>"@domain.dom

Sendmail/Exim on the other hand require '<>'.  If you attempt to use
'' under Sendmail, <user@hostname> will be substituted instead.  Under
Exim, it flat out fails:

  $ echo | /usr/sbin/sendmail -i -f "" -- you@yourdomain.dom
  exim:  - bad address: empty address

We now address this in Util.sendmail() which is where all TMDA mail
is sent from.

This diff will be applied to the main trunk as well.