Looking through the CVS history, I see two more problems that I forgot about.  Courier and some versions of Postfix will refuse an SMTP transmission when MAIL FROM is an empty string.  Also, some old versions of Postfix cannot handle "sendmail -f <>".
So, if you boil it all down, to send mail with a null envelope sender,
the string '<>' should be used in all cases except for the following,
where an empty string should be used instead:

When MAIL_TRANSPORT is 'sendmail' *AND* MAIL_TRANSFER_AGENT is either
'postfix' or 'qmail'.

I really hope this is the last time we have to deal with this
discrepancy.