tmda-ofmipd: Wrap message processing in an exception handler.
This new code catches all exceptions thrown during message processing,
and returns a valid SMTP error response to the client application.
Without this, exceptions (e.g. caused by incorrect syntax in the
outgoing filter) will simply abort the SMTP connection. This confuses
some clients (e.g. Thunderbird), which then think that the message
was successfully sent, when it was actually dropped. Sending an SMTP
error fixes that (arguably, the client should be fixed, but this
change is good in and of itself.)

Notes:
* This fix is untested on TMDA trunk, since I can't upgrade to 1.1.x
  yet, because I need to patch in run-from-stunnel support (soon!)
  However, the same patch has been working on 1.0.2+ for years.
* Perhaps this exception handler should eventually cover a larger
  block of the code (i.e. the whole function its in) to catch all
  errors at any time during the SMTP protocol.