First cut at the "pending queue abstraction" to allow multiple, user-selectable formats for the TMDA pending queue.  More information on this idea at:
http://wiki.tmda.net/TmdaOneDotTwoTodoList#head-9c4fb0cacc9f995ccfa607c65c2538db235638ce

The interface is not set in stone as of yet.  See TMDA/Queue/Queue.py,
and also OriginalQueue.py for an implementation of the current
"original" style TMDA queue.  This code should work identically to
previous versions without any changes to config files.  I've tested 
tmda-filter and tmda-pending and both seem to work fine.

One notable change is that pending messages are no longer referred to
by their filename in the logs and elsewhere.  Previously, you'd see
things like "HOLD pending 1159315215.2974.msg" in the logs.  Now that
no longer makes sense as depending on the format of the queue, the
message may or may not be stored as a flat file.  Thus, we've dropped
the '.msg' suffix, and just refer to the message as "1159315215.2974".
This can be thought of as a "mailid" that uniquely identifies the
message, no matter what kind of pending queue it's stored in.