New features from Lloyd Zusman:
He explains:

I'm using TMDA with the Courier MTA, and all my users get virtual email
directories as implemented in the Courier manner.  On my system, all of
these are subdirectories which live under /var/vmail, and all are owned
by uid=vmail and gid=vmail.  The .tmda and Maildir directories live in
this subdirectory.  For example, for the virtual email address
"vuser@vdomain.com", we have the following items:

  /var/vmail/vuser@vdomain.com          - the top-level email directory
  /var/vmail/vuser@vdomain.com/.tmda    - the associated tmda home
  /var/vmail/vuser@vdomain.com/Maildir  - the associated Maildir

I can make use of the --vhost-script option for tmda-filter and
tmda-ofmipd to cause each of these programs to set HOME to the
appropriate virtual directory, but I am unable to perform the
corresponding function with tmda-pending, which lacks the --vhost-script
option.

Furthermore, the method in http://wiki.tmda.net/VirtualDomainsHowto for
using tmda-pending with virtual domains does not work for my Courier
virtual email setup ... tmda-pending still can't find the pending
directory.

Up until now, I have always manually hacked the TMDA source code to
allow my users to utilize tmda-pending with their Courier virtual mail
directories.  However, I'm getting tired of having to reinstall this
hack every time I upgrade TMDA.

Therefore, I have implemented a patch for tmda-pending which also gives
it a --vhost-script option.  It works the same as it does for
tmda-filter and tmda-ofmipd.  The script specified by this option will
be passed the user name and the domain name as determined with the
following algorithm (which makes use of two other new options:
--vhost-user and --vhost-domain) ...

  user name:   if the --vhost-user option is set, use it

               else if the USER environment variable is set,
               use it

               else if the LOGNAME environment variable is set,
               use it

               else ERROR

  domain name: if the --vhost-domain option is set, use it

               else if socket.gethostbyaddr(socket.gethostname())[0]
               succeeds, use this value

               else ERROR