tmda-ofmipd: Prevent 100% CPU usage when the client disconnects.
This can be reproduced by running the following command, waiting
for the SMTP server signon message, and typing QUIT<enter>.

openssl s_client -connect 127.0.0.1:8027 -crlf

This appears to be a bug in tlslite, because it continually calls
outReadEvent with an empty string.

The fix we apply here is probably the fix that should be applied
to tlslite itself. The fix is however applied to tmda-ofmipd in
order not to require a patched tlslite. Besides, the tlslite
TLSAsyncDispatcherMixIn is duplicated in tmda-ofmipd in order to
add features (STARTTLS support). Possibly, though, this fix should
be applied to the caller of outReadEvent so as to apply to all
tlslite users, not just the asyncore integration. Still, even if
the bug is fixed there, this fix here won't hurt anything.