tmda-ofmipd: A *much* simpler way of making handle_connect work.
Since we implement our own version of TLSAsyncDispatcherMixIn, we can make it call handle_connect directly on the object it should be called on. This removes the need to insert an extra class into the inheritance hierarchy just to hook handle_connect.

This change has been proposed to the tlslite developer, and hopefully will make it into some future release of tlslite.