Fix authentication test failure.
The test itself needed to be updated to use AUTH LOGIN\r\n and put the
username in the next message. This is because AUTH LOGIN <username>\r\n
is indistinguishable from the AUTH LOGIN\r\n case when the username is
empty (which is one of the test cases).

tmda-ofmipd needed to be updated to properly look for None instead of
directly testing objects that could be None or empty strings.